+++ live_view_studio/lib/live_view_studio_web/live/boats_live.ex
@@ -2,6 +2,7 @@ defmodule LiveViewStudioWeb.BoatsLive do
   use LiveViewStudioWeb, :live_view
 
   alias LiveViewStudio.Boats
+  import LiveViewStudioWeb.CustomComponents
 
   def mount(_params, _session, socket) do
     socket =
@@ -23,6 +24,26 @@ defmodule LiveViewStudioWeb.BoatsLive do
       </:legal>
     </.promo>
     <div id="boats">
+      <.filter_form filter={@filter} />
+
+      <div class="boats">
+        <.boat :for={boat <- @boats} boat={boat} />
+      </div>
+    </div>
+
+    <.promo>
+      Hurry, only 3 boats left.
+      <:legal>
+        Excluding weekends.
+      </:legal>
+    </.promo>
+    """
+  end
+
+  attr :filter, :map, required: :true
+
+  def filter_form(assigns) do
+    ~H"""
       <form phx-change="type-change">
         <div class="filters">
           <select name="type">
@@ -46,49 +67,30 @@ defmodule LiveViewStudioWeb.BoatsLive do
           </div>
         </div>
       </form>
-      <div class="boats">
-        <div :for={boat <- @boats} class="boat">
-          <img src={boat.image} />
+    """
+  end
+
+  attr :boat, LiveViewStudio.Boats.Boat, required: :true
+
+  def boat(assigns) do
+    ~H"""
+      <div class="boat">
+          <img src={@boat.image} />
           <div class="content">
             <div class="model">
-              <%= boat.model %>
+              <%= @boat.model %>
             </div>
             <div class="details">
               <span class="price">
-                <%= boat.price %>
+                <%= @boat.price %>
               </span>
               <span class="type">
-                <%= boat.type %>
+                <%= @boat.type %>
               </span>
             </div>
           </div>
         </div>
-      </div>
-    </div>
-
-    <.promo expiration={1}>
-      Hurry, only 3 boats left.
-      <:legal>
-        Excluding weekends.
-      </:legal>
-    </.promo>
-    """
-  end
-
-  def promo(assigns) do
-    ~H"""
-      <div class="promo">
-        <div class="deal">
-          <%= render_slot(@inner_block) %>
-        </div>
-        <div class="expiration">
-          Deal expires in <%= @expiration %>
-        </div>
-        <div class="legal">
-          <%= render_slot(@legal) %>
-        </div>
-      </div>
-    """
+      """
   end
 
   def handle_event("type-change", params, socket) do
